backend/requirements.txt
Django
djangorestframework
djoser
python-dotenv
djangorestframework-simplejwt
django-filter
